CN-4930 Split transactions util into two utils
