+++ app/redux/transactions/transactionsReducer.js
@@ -1,6 +1,6 @@
 import ACTIONS from "../../constants/ACTIONS";
-import transactionsUtil from "../../utils/transactions-util";
 import { mocks } from "../../../dev-settings";
+import transactionMapUtil from "../../utils/transaction-map-util";
 
 /**
  * TODO make it a function add JSDoc & desc for return
@@ -19,11 +19,11 @@ export default function transactionsReducer(state = initialState, action) {
         Object.values(
           require("../../mock-data/transactions.mock").default
         ).forEach(t => {
-          newTransactions[t.id] = transactionsUtil.mapTransaction(t);
+          newTransactions[t.id] = transactionMapUtil.mapTransaction(t);
         });
       } else {
         action.transactions.forEach(t => {
-          newTransactions[t.id] = transactionsUtil.mapTransaction(t);
+          newTransactions[t.id] = transactionMapUtil.mapTransaction(t);
         });
       }
 
@@ -37,7 +37,9 @@ export default function transactionsReducer(state = initialState, action) {
     case ACTIONS.GET_TRANSACTION_DETAILS_SUCCESS:
     case ACTIONS.CANCEL_WITHDRAWAL_TRANSACTION_SUCCESS:
     case ACTIONS.WITHDRAW_CRYPTO_SUCCESS:
-      transactionDetails = transactionsUtil.mapTransaction(action.transaction);
+      transactionDetails = transactionMapUtil.mapTransaction(
+        action.transaction
+      );
       return {
         ...state,
         transactionDetails: { ...transactionDetails },
